Enable support for python to dlopen .so.avx2 libs
Summary:
With this patch python can first try and dlopen .so.avx2 libs if
they exist on the system and the cpu is supported.

This is neccessary as python dlopens solib files and as such standard
glibc hwcaps mechasism can't be used.

Test Plan:
Confirm .so.avx2 libs are loaded on a supported system
Confirm normal .so libs are loaded if .so.avx2 libs don't exist
TODO: ensure .so.avx2 files are ignored on a non-supported system.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.getsol.us/D11518